% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_allowed_values}
\alias{volt_allowed_values}
\title{Allowed values used by VOLT workflows}
\usage{
volt_allowed_values()
}
\value{
Named list of allowed categorical values.
}
\description{
Allowed values used by VOLT workflows
}
\examples{
# Usa questi valori per evitare codifiche incoerenti nei file da importare.
values <- volt_allowed_values()
values$faro_status
values$direction
values$workflow_mode
}
