% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_alternatives}
\alias{volt_alternatives}
\title{Create a VOLT alternative table}
\usage{
volt_alternatives(data)
}
\arguments{
\item{data}{Data frame with \code{alternative_id}, \code{label}, and \code{faro_status}.}
}
\value{
A \code{volt_alternatives} data frame.
}
\description{
Create a VOLT alternative table
}
