% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_apply_import}
\alias{volt_apply_import}
\title{Apply imported data to a VOLT project bundle}
\usage{
volt_apply_import(
  project,
  object,
  target = c("auto", "alternatives", "criteria", "performance")
)
}
\arguments{
\item{project}{VOLT project bundle.}

\item{object}{Imported object.}

\item{target}{One of \code{auto}, \code{alternatives}, \code{criteria}, or \code{performance}.}
}
\value{
Updated project bundle.
}
\description{
Apply imported data to a VOLT project bundle
}
