% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_import_file}
\alias{volt_import_file}
\title{Import a local file for VOLT}
\usage{
volt_import_file(path)
}
\arguments{
\item{path}{Path to a CSV, RDS, RDA, or RData file.}
}
\value{
Imported R object.
}
\description{
Import a local file for VOLT
}
\examples{
# Importa un file locale CSV, RDS, RDA o RData.
path <- volt_example_path("esempio_criteri.csv")
imported <- volt_import_file(path)
imported
}
