% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_indicator_performance}
\alias{volt_indicator_performance}
\title{Create a VOLT atomic indicator performance table}
\usage{
volt_indicator_performance(data)
}
\arguments{
\item{data}{Data frame with \code{alternative_id}, \code{indicator_id}, and \code{value}.}
}
\value{
A \code{volt_indicator_performance} data frame.
}
\description{
Create a VOLT atomic indicator performance table
}
