% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_write_dia_examples}
\alias{volt_write_dia_examples}
\title{Write bundled DIA-style example CSV files}
\usage{
volt_write_dia_examples(path)
}
\arguments{
\item{path}{Output directory.}
}
\value{
Data frame with written files.
}
\description{
Write bundled DIA-style example CSV files
}
