% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_allowed_values}
\alias{volt_allowed_values}
\title{Allowed values used by VOLT workflows}
\usage{
volt_allowed_values()
}
\value{
Named list of allowed categorical values.
}
\description{
Allowed values used by VOLT workflows
}
