% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_apply_ahp_to_clusters}
\alias{volt_apply_ahp_to_clusters}
\title{Apply AHP-derived weights to DIA clusters}
\usage{
volt_apply_ahp_to_clusters(project, ahp)
}
\arguments{
\item{project}{DIA-aware VOLT project bundle.}

\item{ahp}{A \code{volt_ahp_weights} object.}
}
\value{
Updated DIA-aware project bundle.
}
\description{
Apply AHP-derived weights to DIA clusters
}
