% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_example_path}
\alias{volt_example_path}
\title{Locate a bundled VOLT upload example}
\usage{
volt_example_path(file)
}
\arguments{
\item{file}{Example file name.}
}
\value{
Installed file path.
}
\description{
Locate a bundled VOLT upload example
}
\examples{
# Recupera il percorso di un file di esempio installato con il pacchetto.
path <- volt_example_path("esempio_alternative.csv")
path
utils::read.csv(path)
}
