% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_formula_reference}
\alias{volt_formula_reference}
\title{VOLT formula reference table}
\usage{
volt_formula_reference()
}
\value{
Data frame with formula steps used by \code{volt_rank()}.
}
\description{
VOLT formula reference table
}
