% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_open_project}
\alias{volt_open_project}
\title{Open a saved VOLT project bundle}
\usage{
volt_open_project(path)
}
\arguments{
\item{path}{RDS path.}
}
\value{
VOLT project bundle with recalculated ranking.
}
\description{
Open a saved VOLT project bundle
}
