% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_performance}
\alias{volt_performance}
\title{Create a VOLT performance matrix}
\usage{
volt_performance(data)
}
\arguments{
\item{data}{Data frame with \code{alternative_id}, \code{criterion_id}, and \code{value}.}
}
\value{
A \code{volt_performance} data frame.
}
\description{
Create a VOLT performance matrix
}
