% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_save_project}
\alias{volt_save_project}
\title{Save a VOLT project bundle}
\usage{
volt_save_project(project, path)
}
\arguments{
\item{project}{VOLT project bundle.}

\item{path}{Output RDS path.}
}
\value{
The saved path.
}
\description{
Save a VOLT project bundle
}
