% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_upload_examples}
\alias{volt_upload_examples}
\title{List bundled VOLT upload examples}
\usage{
volt_upload_examples()
}
\value{
Data frame with bundled example files and their purpose.
}
\description{
List bundled VOLT upload examples
}
