% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{volt_upload_guide}
\alias{volt_upload_guide}
\title{Explain VOLT data imports}
\usage{
volt_upload_guide()
}
\value{
Data frame describing accepted upload types, required columns, and
operational purpose.
}
\description{
Explain VOLT data imports
}
